chore: update nats-server go module
NATS Server releases v2.11.1 and v2.10.27 include a fix
for a security vulnerability (CVE):
https://advisories.nats.io/CVE/secnote-2025-01.txt

This PR updates to a patched version that is not vulnerable.

If applicable to this repo, we should also cut a release after merging.

This is a batch change created with https://github.com/lindell/multi-gitter
```
  go get github.com/nats-io/nats-server/v2@patch
  go get toolchain@none
  go mod tidy
```

All PRs in this batch:
https://github.com/search?q=label%3Aabozhenko_multigitter_bump_nats_server_cve_2025_04+state%3Aopen&type=pullrequests

Signed-off-by: Phil Pennock <pdp@synadia.com>